requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.3
python-dateutil==2.8.1
requests==2.25.1
setuptools~=57.0.0
termcolor==1.1.0
xhtml2pdf==0.2.5
